hynix_dashboard/requirements.txt
pandas
numpy
gdown
plotly
scipy
streamlit-plotly-events
